Separating chart component from page file
